feat:authorization

+++ Modules/Users/Http/Controllers/RolesController.php
@@ -9,6 +9,7 @@ use App\Http\Controllers\AppBaseController;
 use Modules\Users\Transformers\RoleResource;
 use Modules\Users\Repositories\RoleRepository;
 use Symfony\Component\HttpFoundation\Response;
+use Illuminate\Auth\Access\AuthorizationException;
 use Modules\Users\Http\Requests\CreateRoleRequest;
 use Modules\Users\Http\Requests\UpdateRoleRequest;
 
@@ -26,10 +27,14 @@ class RolesController extends AppBaseController
 
     /**
      * Display a listing of the resource.
+     * @param Request $request
      * @return Response
+     * @throws AuthorizationException
      */
     public function index(Request $request): Response
     {
+        $this->authorize('view', Role::class);
+
         $roles = $this->roleRepository->all(
             Arr::except($request->all(), ['limit', 'skip']),
             $request->get('skip'),
@@ -43,9 +48,12 @@ class RolesController extends AppBaseController
      * Store a newly created resource in storage.
      * @param CreateRoleRequest $request
      * @return Response
+     * @throws AuthorizationException
      */
-    public function store(CreateRoleRequest $request): Response
+    public function store(CreateRoleRequest $request)
     {
+        $this->authorize('create', Role::class);
+
         $role = $this->roleRepository->create($request->all());
 
         return $this->sendResponse(new RoleResource($role), 'Role successfully created');
@@ -55,9 +63,12 @@ class RolesController extends AppBaseController
      * Show the specified resource.
      * @param int $id
      * @return Response
+     * @throws AuthorizationException
      */
     public function show(int $id): Response
     {
+        $this->authorize('show', Role::class);
+
         $role = $this->roleRepository->find($id);
 
         if (empty($role))
@@ -71,9 +82,12 @@ class RolesController extends AppBaseController
      * @param UpdateRoleRequest $request
      * @param Role $role
      * @return Response
+     * @throws AuthorizationException
      */
     public function update(UpdateRoleRequest $request, Role $role): Response
     {
+        $this->authorize('update', Role::class);
+
         $role = $this->roleRepository->update($request->all(), $role->id);
 
         return $this->sendResponse(new RoleResource($role), 'Role updated successfully');
@@ -83,9 +97,12 @@ class RolesController extends AppBaseController
      * Remove the specified resource from storage.
      * @param Role $role
      * @return Response
+     * @throws AuthorizationException
      */
     public function destroy(Role $role): Response
     {
+        $this->authorize('delete', Role::class);
+
         $role->delete();
 
         return $this->sendResponse(null, 'Role successfully deleted');
